requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==4.5.4
psycopg2-binary==2.9.7
pywin32==306; sys_platform == "win32"
requests==2.31.0
Pillow==10.0.1
python-docx==1.1.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
drf-yasg==1.21.4
django-extensions==3.2.3